Extract getting newsletter stats out into a custom hook
