+++ plugins/newsletter-status/newsletterStatusPanel.tsx
@@ -1,48 +1,20 @@
-import React, { useCallback, useEffect, useState } from '@wordpress/element';
+import React from '@wordpress/element';
 import { PluginDocumentSettingPanel } from '@wordpress/edit-post';
 import { __ } from '@wordpress/i18n';
 import { select } from '@wordpress/data';
-import apiFetch from '@wordpress/api-fetch';
 import { Button } from '@wordpress/components';
 import NewsletterSpinner from '@/components/newsletterSpinner';
-
-interface Status {
-  Bounced?: number;
-  Clicks?: number;
-  Forwards?: number;
-  Likes?: number;
-  Mentions?: number;
-  Name?: string;
-  Recipients?: number;
-  SpamComplaints?: number;
-  Status?: string;
-  TotalOpened?: number;
-  UniqueOpened?: number;
-  Unsubscribed?: number;
-  WebVersionTextURL?: string;
-  WebVersionURL?: string;
-  WorldviewURL?: string;
-}
+import useNewsletterStatus from '@/hooks/useNewsletterStatus';
 
 export default function NewsletterStatusPanel() {
   // @ts-ignore
   const postId = select('core/editor').getCurrentPostId();
-
-  const [status, setStatus] = useState<Status>({});
-  const [fetching, setFetching] = useState(false);
-
-  const fetchStatus = useCallback(async () => {
-    setFetching(true);
-    const res = await apiFetch({
-      path: `/wp-newsletter-builder/v1/status/${postId}`,
-    });
-    setStatus(res as Status);
-    setFetching(false);
-  }, [postId]);
-
-  useEffect(() => {
-    fetchStatus();
-  }, [fetchStatus]);
+  const {
+    status,
+    fetching,
+    fetchStatus,
+    validStatus,
+  } = useNewsletterStatus(postId);
 
   const {
     Status: statusString = '',
@@ -52,7 +24,7 @@ export default function NewsletterStatusPanel() {
     UniqueOpened = '',
   } = status;
 
-  if (!statusString || !Name) {
+  if (!validStatus && !fetching) {
     return (
       <PluginDocumentSettingPanel
         name="rubric-selection"
@@ -77,7 +49,7 @@ export default function NewsletterStatusPanel() {
       name="rubric-selection"
       title={__('Newsletter Status', 'wp-newsletter-builder')}
     >
-      {status ? (
+      {validStatus ? (
         <>
           <dl>
             <dt>
